requirements.txt
Django==4.0.1
python-dotenv==0.19.2
djangorestframework==3.13.1
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.0.0
thefuzz[speedup]==0.19.0
